backend/requirements.txt
requests>=2.31.0
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
python-dotenv>=0.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dateutil>=2.8.0
openai>=1.0.0
